Model with major problems
I changed around the "hear_data_clean" because it didn't look like it was doing what it was supposed to do exactly (not sure if correct). There are some major problems with the modle mainly that the ccuracy is so low and I'm not sure how to get it higher.